simple while two vars

+++ graphs/mathops/add_n/add_n_save_model.py
@@ -9,9 +9,9 @@ ops = ["add"
     , "add_n"]
 
 my_feed_dict = {}
-in0 = tf.placeholder("float", [3, 3], name="input_0")
+in0 = tf.placeholder("float64", [3, 3], name="input_0")
 my_feed_dict[in0] = get_input("input_0")
-in1 = tf.placeholder("float", [3, 3], name="input_1")
+in1 = tf.placeholder("float64", [3, 3], name="input_1")
 my_feed_dict[in1] = get_input("input_1")
 k0 = tf.Variable(tf.random_normal([3, 3]), name="in0", dtype=tf.float32)
