+++ graphs/mathops/single_arg_ops/single_arg_ops.py
@@ -37,9 +37,9 @@ model_name = "transform_0"
 save_dir = model_name
 
 my_feed_dict = {}
-in0 = tf.placeholder("float", [3, 3], name="input_0")
-in1 = tf.placeholder("float", [3, 3], name="input_1")
-k0 = tf.Variable(tf.random_normal([3, 3]), name="in0", dtype=tf.float32)
+in0 = tf.placeholder("float64", [3, 3], name="input_0")
+in1 = tf.placeholder("float64", [3, 3], name="input_1")
+k0 = tf.Variable(tf.random_normal([3, 3]), name="in0")
 my_feed_dict[in0] = get_input("input_0")
 my_feed_dict[in1] = get_input("input_1")
